♻️ ツーリング編集をモーダル化・削除を一覧に移動
- 編集ページを廃止し、TouringEditForm / TouringEditModalコンポーネントに分割
- 詳細ページのツーリング情報カード右上にペンアイコンで編集モーダルを表示
- モーダルにSP全画面・PC中央寄せ対応とXボタン(閉じる)を追加
- 削除機能を編集フォームから取り除き、一覧のゴミ箱ボタンに移動
