+++ apps/web/components/touring/TouringEditForm.module.css
@@ -1,0 +1,6 @@
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin-bottom: 1.5rem;
+}
