+++ apps/web/components/touring/TouringListSection.tsx
@@ -13,14 +13,14 @@ export interface TouringListSectionProps {
   tourings: ApiResponseTouringDetail[]
 
   /**
-   * ツーリング履歴編集時のコールバック
-   */
-  onEdit: (touringId: string) => void
-
-  /**
    * ツーリング履歴詳細表示時のコールバック
    */
   onDetail?: (touringId: string) => void
+
+  /**
+   * ツーリング履歴削除時のコールバック
+   */
+  onDelete?: (touringId: string) => void
 
   /**
    * ツーリング履歴登録ボタンクリック時のコールバック
@@ -30,8 +30,8 @@ export interface TouringListSectionProps {
 
 export const TouringListSection = ({
   tourings,
-  onEdit,
   onDetail,
+  onDelete,
   onRegister,
 }: TouringListSectionProps) => {
   return (
@@ -49,8 +49,8 @@ export const TouringListSection = ({
             <TouringListItem
               key={touring.touringId}
               touring={touring}
-              onEdit={onEdit}
               onDetail={onDetail}
+              onDelete={onDelete}
             />
           ))}
         </div>
